requirement.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
dictop==0.1.4
Django==2.1.7
fpdf==1.7.2
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.10
grpcio==1.20.1
idna==2.8
joblib==0.13.2
nltk==3.4.1
numpy==1.16.3
pdf2image==1.5.4
Pillow==6.0.0
pkg-resources==0.0.0
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2018.9
requests==2.22.0
rsa==4.0
scikit-learn==0.21.1
scipy==1.3.0
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
urllib3==1.25.2
